Added fake login servlet, update sdk to 1.8.2
The fake login servlet is for debugging when offline and only accessible
to admins under the GAE login system. Also, the jar files for the
appengine sdk were updated to 1.8.2.

Also: isProduction property in Utils, user.toString() and used that for
logging, fixed issue with + in session id by using UriBuilder,
miscellaneous changes