Updated Java libraries Every Java library that we depend on has been updated to the latest version. java_deps.md now has a nice synthesis of each of the projects we depend on, and their jars. * App Engine SDK to 1.8.9 * Objectify to 4.0rc2 * Jersey to 2.5.1 * Jackson to 2.2.3 * Guava to 16.0 (dependency of Objectify, Jersey)
Jersey 2.x brought along a lot of new jars that were previously combined in the Jersey 1.17 bundle. I didn't add all of the jars, so that may be necessary as time goes on.

With the addition of Jackson as a JSON provider for JAX-RS, we no longer have to manually interface ObjectMapper.
Also, just for good measure, we now simply return the object rather than a Response object, as Jackson will process it.
The downside to this change is that the JSON padding no longer exists. Hopefully this isn't too much of a problem, as it seems like it will be difficult to fix.